moved react tests and stories into react folders, excluded Jest pattern to match only react tests
