+++ .storybook/config.js
@@ -2,7 +2,7 @@ import { configure } from '@storybook/react';
 import './storybook.scss';
 
 function loadStories() {
-  require('../stories');
+  require('../stories/react');
 }
 
 configure(loadStories, module);
