+++ jest.config.js
@@ -8,20 +8,18 @@ module.exports = {
 	},
 	transform: {
 		'^.+\\.js$': 'babel-jest',
-		'^.+\\.html$': '<rootDir>/test/utils/htmlLoader.js'
+		'^.+\\.html$': '<rootDir>/test/react/utils/htmlLoader.js'
 	},
 	moduleNameMapper: {
-		'^.+\\.svg$': '<rootDir>/test/utils/svgMock.js'
+		'^.+\\.svg$': '<rootDir>/test/react/utils/svgMock.js'
 	},
 	coveragePathIgnorePatterns: [
 		'<rootDir>/src/react/index.js',
 		'<rootDir>/src/index.js'
 	],
+	testRegex: '/test/react/.*.(spec|test)\\.jsx?$',
 	testPathIgnorePatterns: [
 		'<rootDir>/node_modules/',
-        '<rootDir>/webpack/webpack.test.js',
-        '<rootDir>/src/angular/',
-        '<rootDir>/lib/angular/'
 	],
 	collectCoverageFrom: [
 		'src/**/*.{js,jsx}'
