+++ stories/react/buttons/LinkButtons.stories.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import Examples from '../utils/Examples.js';
 
-import ReactButton from '../../src/react/Button.js';
+import ReactButton from '../../../src/react/Button.js';
 
-import HTMLButtonLinkPrimary from '../../components/button/button-link-primary.html';
-import HTMLButtonLinkPrimaryDisabled from '../../components/button/button-link-primary-disabled.html';
-import HTMLButtonLinkPrimaryWithIcon from '../../components/button/button-link-primary-with-icon.html';
-import HTMLButtonLinkPrimaryWithIconDisabled from '../../components/button/button-link-primary-with-icon-disabled.html';
+import HTMLButtonLinkPrimary from '../../../components/button/button-link-primary.html';
+import HTMLButtonLinkPrimaryDisabled from '../../../components/button/button-link-primary-disabled.html';
+import HTMLButtonLinkPrimaryWithIcon from '../../../components/button/button-link-primary-with-icon.html';
+import HTMLButtonLinkPrimaryWithIconDisabled from '../../../components/button/button-link-primary-with-icon-disabled.html';
 
-import HTMLButtonLinkSecondary from '../../components/button/button-link-secondary.html';
-import HTMLButtonLinkSecondaryDisabled from '../../components/button/button-link-secondary-disabled.html';
-import HTMLButtonLinkSecondaryWithIcon from '../../components/button/button-link-secondary-with-icon.html';
-import HTMLButtonLinkSecondaryWithIconDisabled from '../../components/button/button-link-secondary-with-icon-disabled.html';
+import HTMLButtonLinkSecondary from '../../../components/button/button-link-secondary.html';
+import HTMLButtonLinkSecondaryDisabled from '../../../components/button/button-link-secondary-disabled.html';
+import HTMLButtonLinkSecondaryWithIcon from '../../../components/button/button-link-secondary-with-icon.html';
+import HTMLButtonLinkSecondaryWithIconDisabled from '../../../components/button/button-link-secondary-with-icon-disabled.html';
 
 
 let examples = {
