+++ stories/react/utils/InsertSVGIcons.js
