+++ test/react/Tile.spec.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import Tile from '../src/react/Tile.js';
-import TileInfo from '../src/react/TileInfo.js';
-import TileInfoLine from '../src/react/TileInfoLine.js';
-import TileFooter from '../src/react/TileFooter.js';
-import TileFooterCell from '../src/react/TileFooterCell.js';
+import Tile from '../../src/react/Tile.js';
+import TileInfo from '../../src/react/TileInfo.js';
+import TileInfoLine from '../../src/react/TileInfoLine.js';
+import TileFooter from '../../src/react/TileFooter.js';
+import TileFooterCell from '../../src/react/TileFooterCell.js';
 
 import renderer from 'react-test-renderer';
