+++ test/react/utils/htmlLoader.js
