+++ test/react/utils/svgMock.js
